refactor: streamline memory usage calculations and remove unused
