+++ website-nextjs/src/components/admin/ResultsSections.tsx
@@ -7,8 +7,9 @@ import { roundNumber } from "@/utils/number";
 import Popup from "reactjs-popup";
 import { IFilterState, IResultState } from "@/types/state";
 import ResultsSectionsTitle from "./home/ResultsTitle";
-import { SgmMode, TIMEOUT_VALUES } from "@/constants/filter";
+import { SgmMode } from "@/constants/filter";
 import { extractNumberFromFormattedString } from "@/utils/string";
+import { getMaxMemoryUsage } from "@/utils/results";
 
 type ColumnType = {
   name: string;
@@ -41,6 +42,10 @@ const ResultsSection = ({ timeout }: ResultsSectionProps) => {
     },
   ).filter((result) => result.timeout === timeout);
 
+  const rawMetaData = useSelector((state: { results: IResultState }) => {
+    return state.results.rawMetaData;
+  });
+
   const availableSolvers = useSelector((state: { results: IResultState }) => {
     return state.results.availableSolvers;
   });
@@ -57,8 +62,6 @@ const ResultsSection = ({ timeout }: ResultsSectionProps) => {
   });
 
   const benchmarkResults = useMemo(() => {
-    const maxMemoryUsage =
-      timeout === TIMEOUT_VALUES.SHORT ? 7 * 1024 : 62 * 1024;
     switch (sgmMode) {
       case SgmMode.ONLY_ON_INTERSECTION_OF_SOLVED_BENCHMARKS:
         const benchmarkSuccessMap = new Map<string, number>();
@@ -90,7 +93,7 @@ const ResultsSection = ({ timeout }: ResultsSectionProps) => {
             result.status !== "ok" ? result.timeout * xFactor : result.runtime,
           memoryUsage:
             result.status !== "ok"
-              ? maxMemoryUsage * xFactor
+              ? getMaxMemoryUsage(result, rawMetaData) * xFactor
               : result.memoryUsage,
         }));
       case SgmMode.COMPUTE_SGM_USING_TO_VALUES:
@@ -98,7 +101,9 @@ const ResultsSection = ({ timeout }: ResultsSectionProps) => {
           ...result,
           runtime: result.status !== "ok" ? result.timeout : result.runtime,
           memoryUsage:
-            result.status !== "ok" ? maxMemoryUsage : result.memoryUsage,
+            result.status !== "ok"
+              ? getMaxMemoryUsage(result, rawMetaData)
+              : result.memoryUsage,
         }));
       default:
         return benchmarkLatestResults;
