+++ website-nextjs/src/constants/index.ts
@@ -8,8 +8,6 @@ export enum ProblemSize {
 
 export const MaxRunTime = 36000;
 
-export const MaxMemoryUsage = 7 * 1024;
-
 export enum Technique {
   MILP = "MILP",
   LP = "LP",
